moved test_listener_ip_port back to its own file
