+++ ucp/_libs/tests/test_server_client.py
@@ -11,21 +11,6 @@ from ucp._libs.utils_test import blocking_recv, blocking_send
 mp = mp.get_context("spawn")
 
 
-def test_listener_ip_port():
-    ctx = ucx_api.UCXContext()
-    worker = ucx_api.UCXWorker(ctx)
-
-    def _listener_handler(conn_request):
-        pass
-
-    listener = ucx_api.UCXListener(worker=worker, port=0, cb_func=_listener_handler)
-
-    assert isinstance(listener.ip, str) and listener.ip
-    assert (
-        isinstance(listener.port, int) and listener.port >= 0 and listener.port <= 65535
-    )
-
-
 def _echo_server(queue, msg_size):
     """Server that send received message back to the client
